+++ src/main/java/ch/saunah/saunahbackend/dto/SaunaResponse.java
@@ -10,7 +10,7 @@ public class SaunaResponse {
     private String name;
     private String description;
     private boolean isMobile;
-    private int price;
+    private float price;
     private int maxTemp;
     private int numberOfPeople;
     private String location;
@@ -53,7 +53,7 @@ public class SaunaResponse {
         return isMobile;
     }
 
-    public int getPrice() {
+    public float getPrice() {
         return price;
     }
